requirements.txt
fastapi
uvicorn
httpx
python-dotenv
openai
jinja2
requests